Methylation path on chosen gene.

Function plot_methylation_path visualizes a chosen gene with marked CpG probes. Y axis describes methylation level. X axis describes a location of the probe on the chromosome. Horizontal lines show the mean methylation level for each Island in a division to groups. Groups are defined by colors. Large dots symbolize means of methylation level for CpG probes, small dots symbolize methylation levels for each observation. Also, we can exact the line corresponding to a gene. In this case, we see what are the locations of probes on a gene in HG18 coordinates. We can as well draw locations of CpG islands.

plot_methylation_path(data, condition, gene, show.gene = FALSE,
  observ = FALSE, islands = TRUE, title = TRUE, genom.data = NULL,
  mean.size = 4, observ.size = 0.5, ...)

Arguments

data
Data frame containing values from methylation: columns corespond to CpG probes, rows to samples.

condition
Vector of levels corresponding to order of samples in data.

gene
Name of chosen gene.

show.gene
Logical. If TRUE arrows corresopnding to gene will be drawn.

observ
Logical. If TRUE dots corresponding to CpG probes will be drawn.

islands
Logical. If TRUE line corresopnding to islands should be drawn.

title
Logical. If TRUE title saying what gene we visualise will be added.

genom.data
Data frame which contains information about CpG probes and corresponding genes, by default in our package we use illumina_humanmethylation_27_data. New dataset should contain 5 columns with: 1) CpG probe names, 2) CpG probe locations, 3) gene names, 4) logical value if there is a CpG island, 5) location of island.

mean.size
size of mean dots

observ.size
size of observation dots

...
Other parameters.

Value

Object of class ggplot containing visualisation of methylation on gene.

See also

Examples

# NOT RUN {
library(MLExpRessoData)
condition_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA", "other")
plot_methylation_path(BRCA_met, condition_met, "BRCA2")
# }